separate functions, this will finally fix everything
